Sort Thunderbird contacts by PopularityIndex.
Do doesn't seem to pay attention to ChildrenOfItem when selecting the main ContactItem,
which makes for rather ugly user experience when dealing with multuple emails
for the same contact. This change tries to alleviate this by sorting emails by Thunderbird's
PopularityIndex, i.e., the most frequently used email for a Contact will become
the default one sooner or later.