Rhythmbox: Index playlists and songs; add DBus dependency
With this commit, the Rhythmbox plugin not only indexes artists and albums,
but also playlists and individual songs.
Because rhythmbox-client doesn't allow for switching playlists,
the MPRIS (Media Player Remote Interfacing Specification) DBus interface
is used for this purpose. It fetches the playlist names when the index is
refreshed, and the playlist IDs when the user selects the Play action on
a playlist.
The playlist IDs are not cached because they are non-persistent,
so restarting Rhythmbox would break the functionality until the next refresh.